move map component to webpack
